Reconstruct code.
