--- demo/RecyclerViewDemo/app/src/main/java/com/example/leopeng/recyclerviewdemo/MainActivity.java
@@ -1,536 +1,0 @@
-package com.example.leopeng.recyclerviewdemo;
-
-import android.app.ProgressDialog;
-import android.content.ContentValues;
-import android.content.DialogInterface;
-import android.content.Intent;
-import android.database.Cursor;
-import android.database.sqlite.SQLiteDatabase;
-import android.os.AsyncTask;
-import android.support.v4.view.MenuItemCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.SearchView;
-import android.support.v7.widget.Toolbar;
-import android.util.Log;
-import android.util.LruCache;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import android.view.inputmethod.InputMethodManager;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.EditText;
-import android.widget.ListView;
-import android.widget.Toast;
-
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.net.UnknownHostException;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-import java.util.Random;
-
-public class MainActivity extends AppCompatActivity {
-
-    private String username;
-    private String lowercaseUsername;
-    private ProgressDialog progressDialog;
-    private EditText editText;
-    private ViewGroup backgroudView;
-    private ListView searchHistoryListView;
-    public final static String BOOKJSONKEY = "BOOKJSON";
-    public final static String USERNAMEKEY = "USERNAMEKEY";
-    public final static String FIRSTLOADKEY = "FIRSTLOAD";
-    public final static String SEARCHKEY = "SEARCH";
-    public final static String MAINACTIVITYTAG = "MainActivity";
-
-    private final LruCache<String, String> lruCache;
-    private final static int cacheSize = 4 * 1024 * 1024;
-    private Intent intent;
-    private SearchHistory.SearchHistoryDBHelper searchHistoryDBHelper;
-    private List<String> usernameList;
-    private ArrayAdapter adapter;
-
-    public MainActivity() {
-        lruCache = new LruCache<>(cacheSize);
-    }
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.myToolbar);
-        setSupportActionBar(toolbar);
-
-        editText = (EditText) findViewById(R.id.username);
-        backgroudView = (ViewGroup) findViewById(R.id.backgroundView);
-        searchHistoryListView = (ListView) findViewById(R.id.searchHistory);
-        intent = new Intent(this, RecyclerViewActivity.class);
-        searchHistoryDBHelper = new SearchHistory.SearchHistoryDBHelper(this);
-        final InputMethodManager imm = (InputMethodManager) getSystemService(INPUT_METHOD_SERVICE);
-
-        backgroudView.setOnFocusChangeListener(new View.OnFocusChangeListener() {
-            @Override
-            public void onFocusChange(View v, boolean hasFocus) {
-                if (hasFocus) {
-                   imm.hideSoftInputFromWindow(v.getWindowToken(), 0);
-                }
-            }
-        });
-
-        editText.setOnFocusChangeListener(new View.OnFocusChangeListener() {
-            @Override
-            public void onFocusChange(View v, boolean hasFocus) {
-                if (hasFocus) {
-                    searchHistoryListView.setVisibility(View.VISIBLE);
-                } else {
-                    imm.hideSoftInputFromWindow(v.getWindowToken(), 0);
-                    getInfoFromDB();
-                    adapter.notifyDataSetChanged();
-                    searchHistoryListView.setVisibility(View.INVISIBLE);
-                }
-            }
-        });
-
-        usernameList = new ArrayList<>();
-        adapter = new ArrayAdapter<String>(this, R.layout.username_item, usernameList);
-        searchHistoryListView.setAdapter(adapter);
-        searchHistoryListView.setDividerHeight(2);
-
-        searchHistoryListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                username = usernameList.get(position);
-                editText.setText(username);
-                jumpToBookList();
-            }
-        });
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        getInfoFromDB();
-        adapter.notifyDataSetChanged();
-    }
-
-    @Override
-    protected void onDestroy() {
-        searchHistoryDBHelper.close();
-        super.onDestroy();
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.common_menu, menu);
-
-        MenuItem searchItem = menu.findItem(R.id.searchButton);
-        SearchView searchView =  (SearchView) MenuItemCompat.getActionView(searchItem);
-
-        searchView.setQueryHint("Search Books...");
-
-        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-            @Override
-            public boolean onQueryTextSubmit(String query) {
-                Intent intent = new Intent(MainActivity.this, RecyclerViewActivity.class);
-                intent.putExtra(SEARCHKEY, query);
-                startActivity(intent);
-                return true;
-            }
-
-            @Override
-            public boolean onQueryTextChange(String newText) {
-                return false;
-            }
-        });
-
-        return true;
-    }
-
-    private long putInfoIntoDB() {
-        SQLiteDatabase db = searchHistoryDBHelper.getWritableDatabase();
-
-        ContentValues values = new ContentValues();
-        values.put(SearchHistory.SearchHistoryTable.COLUMN_NAME_SEARCH_WORD, lowercaseUsername);
-
-        Date now = new Date();
-        values.put(SearchHistory.SearchHistoryTable.COLUMN_NAME_UPDATED_AT, now.getTime());
-
-        return db.insert(SearchHistory.SearchHistoryTable.TABLE_NAME, null, values);
-    }
-
-    private boolean updateInfoIntoDB(String username) {
-        SQLiteDatabase db = searchHistoryDBHelper.getWritableDatabase();
-
-        ContentValues values = new ContentValues();
-
-        Date now = new Date();
-        values.put(SearchHistory.SearchHistoryTable.COLUMN_NAME_UPDATED_AT, now.getTime());
-
-        String selection = SearchHistory.SearchHistoryTable.COLUMN_NAME_SEARCH_WORD + " LIKE ?";
-        String[] selectionArgs = { username };
-
-        int count = db.update(
-                SearchHistory.SearchHistoryTable.TABLE_NAME,
-                values,
-                selection,
-                selectionArgs
-        );
-
-        return count > 0 ;
-    }
-
-    private boolean isExistInfo() {
-        SQLiteDatabase db = searchHistoryDBHelper.getReadableDatabase();
-        String[] projections = {SearchHistory.SearchHistoryTable._ID};
-
-        String selection = SearchHistory.SearchHistoryTable.COLUMN_NAME_SEARCH_WORD + " = ?";
-        String[] selectionArgs = {lowercaseUsername};
-
-        Cursor cursor = db.query(
-                SearchHistory.SearchHistoryTable.TABLE_NAME,
-                projections,
-                selection,
-                selectionArgs,
-                null,
-                null,
-                ""
-        );
-
-        boolean res = cursor.getCount() > 0;
-        cursor.close();
-        return res ;
-    }
-
-    private void getInfoFromDB() {
-        SQLiteDatabase db = searchHistoryDBHelper.getReadableDatabase();
-        String[] projections = {
-                SearchHistory.SearchHistoryTable._ID,
-                SearchHistory.SearchHistoryTable.COLUMN_NAME_SEARCH_WORD,
-                SearchHistory.SearchHistoryTable.COLUMN_NAME_UPDATED_AT
-        };
-
-        String sortOrder = SearchHistory.SearchHistoryTable.COLUMN_NAME_UPDATED_AT + " DESC";
-        String limits = "5";
-
-        Cursor cursor = db.query(
-                SearchHistory.SearchHistoryTable.TABLE_NAME,
-                projections,
-                null,
-                null,
-                null,
-                null,
-                sortOrder,
-                limits
-        );
-
-        if (usernameList == null) {
-            usernameList = new ArrayList<>();
-        }
-
-        usernameList.clear();
-
-        while (cursor.moveToNext()) {
-            String name = cursor.getString(cursor.getColumnIndexOrThrow(SearchHistory.SearchHistoryTable.COLUMN_NAME_SEARCH_WORD));
-            usernameList.add(name);
-        }
-
-        for (String username : usernameList) {
-            Log.d(MAINACTIVITYTAG, "username: " + username);
-        }
-
-        if (usernameList.isEmpty()) {
-            Log.d(MAINACTIVITYTAG, "No username stored.");
-        }
-
-        cursor.close();
-    }
-
-
-    public void jsonTask(View view) {
-        username = editText.getText().toString();
-        jumpToBookList();
-    }
-
-    private void jumpToBookList() {
-        lowercaseUsername = username.toLowerCase();
-
-        Log.d(MAINACTIVITYTAG, "Username: " + username);
-        Log.d(MAINACTIVITYTAG, "LowerCaseUsername: " + lowercaseUsername);
-
-        if (lowercaseUsername.trim().isEmpty()) {
-            Toast.makeText(this, "Please enter username.", Toast.LENGTH_SHORT).show();
-            return;
-        }
-
-        if (!isExistInfo()) {
-            putInfoIntoDB();
-        } else {
-            updateInfoIntoDB(lowercaseUsername);
-        }
-
-        Intent intent = new Intent(MainActivity.this, RecyclerViewActivity.class);
-        intent.putExtra(USERNAMEKEY, username);
-
-        String jsonString = getCacheJSONString(lowercaseUsername);
-        if (jsonString != null && !jsonString.isEmpty()) {
-            intent.putExtra(BOOKJSONKEY, jsonString);
-            Log.d(MAINACTIVITYTAG, "Hit disk cache!");
-        }
-
-        startActivity(intent);
-
-//        synchronized (lruCache) {
-//            // Hit memory cache
-//            String jsonString = lruCache.get(lowercaseUsername);
-//            if (jsonString != null && !jsonString.isEmpty()) {
-//                intent.putExtra(BOOKJSONKEY, lruCache.get(lowercaseUsername));
-//                intent.putExtra(USERNAMEKEY, username);
-//                intent.putExtra(FIRSTLOADKEY, false);
-//                Log.d(MAINACTIVITYTAG, "Hit memory cache!");
-//                startActivity(intent);
-//            } else {
-//                // Hit disk cache
-//                jsonString = getCacheJSONString(lowercaseUsername);
-//                if (jsonString != null && !jsonString.isEmpty()) {
-//                    intent.putExtra(BOOKJSONKEY, jsonString);
-//                    intent.putExtra(USERNAMEKEY, username);
-//                    intent.putExtra(FIRSTLOADKEY, false);
-//                    Log.d(MAINACTIVITYTAG, "Hit disk cache!");
-//                    startActivity(intent);
-//                } else {
-//                    String url = BookRequest.getUserCollectionsURL(username);
-//                    Log.d(MAINACTIVITYTAG, url);
-//                    new CommonJsonTask(MainActivity.this).executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR, url);
-//                }
-//            }
-//        }
-    }
-
-    private String getCacheJSONString(String fileName) {
-        File file = null;
-        BufferedReader reader = null;
-        try {
-            file = new File(getCacheDir(), fileName);
-            reader = new BufferedReader(new InputStreamReader(new FileInputStream(file)));
-            String line;
-            StringBuffer buffer = new StringBuffer();
-            while ((line = reader.readLine()) != null) {
-                buffer.append(line);
-            }
-            return buffer.toString();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        return null;
-    }
-
-    private class JsonTask extends AsyncTask<String, Integer, String> {
-
-        private volatile boolean running = true;
-        private volatile boolean network = true;
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-
-            progressDialog = new ProgressDialog(MainActivity.this);
-            progressDialog.setMessage("Please wait!");
-            progressDialog.setCancelable(true);
-            progressDialog.setOnCancelListener(new DialogInterface.OnCancelListener() {
-                @Override
-                public void onCancel(DialogInterface dialog) {
-                    cancel(true);
-                }
-            });
-            progressDialog.show();
-        }
-
-        @Override
-        protected String doInBackground(String... params) {
-            HttpURLConnection connection = null;
-            BufferedReader reader = null;
-
-            try {
-                URL url = new URL(params[0]);
-                connection = (HttpURLConnection) url.openConnection();
-                connection.connect();
-
-                int i = 0;
-                Random random = new Random();
-                int stopTime = random.nextInt(35);
-                while (i <= stopTime) {
-                    try {
-                        Thread.sleep(15);
-                        publishProgress(i);
-                        i++;
-                    }
-                    catch (Exception e) {
-                        Log.i(MAINACTIVITYTAG, e.getMessage());
-                    }
-                }
-                Long timeStart = System.currentTimeMillis();
-                Log.d(MAINACTIVITYTAG, "Reader begin! ");
-
-                InputStream stream = connection.getInputStream();
-                reader = new BufferedReader(new InputStreamReader(stream));
-
-                Log.d(MAINACTIVITYTAG, "Reader finish! " + (System.currentTimeMillis() - timeStart));
-
-                StringBuffer buffer = new StringBuffer();
-                String line = "";
-
-                while ((line = reader.readLine()) != null) {
-                    buffer.append(line + "\n");
-                    Log.d("Response: ", "> " + line);
-                }
-
-                // Store to memory cache
-                synchronized (lruCache) {
-                    if (lruCache.get(lowercaseUsername) == null && !buffer.toString().isEmpty() && isTotalNotZero(buffer.toString())) {
-                        lruCache.put(lowercaseUsername, buffer.toString());
-                    }
-                }
-
-                // Store to temp file
-                if (!buffer.toString().isEmpty() && isTotalNotZero(buffer.toString())) {
-                    Log.d(MAINACTIVITYTAG, buffer.toString());
-                    storeFile(lowercaseUsername, buffer.toString());
-                }
-
-                while (i <= 50) {
-                    try {
-                        Thread.sleep(25);
-                        publishProgress(i);
-                        i++;
-                    }
-                    catch (Exception e) {
-                        Log.i(MAINACTIVITYTAG, e.getMessage());
-                    }
-                }
-
-                return buffer.toString();
-
-            } catch (MalformedURLException e) {
-                e.printStackTrace();
-            } catch (UnknownHostException e) {
-                if (progressDialog.isShowing()) {
-                    progressDialog.dismiss();
-                }
-                network = false;
-                if (running) {
-                    runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            Toast.makeText(MainActivity.this, "Please check your network setting.", Toast.LENGTH_LONG).show();
-                        }
-                    });
-                }
-                e.printStackTrace();
-            } catch (IOException e) {
-                e.printStackTrace();
-            } finally{
-                if (connection != null) {
-                    connection.disconnect();
-                }
-                try {
-                    if (reader != null) {
-                        reader.close();
-                    }
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(String s) {
-            super.onPostExecute(s);
-            if (progressDialog.isShowing()) {
-                progressDialog.dismiss();
-            }
-
-            intent.putExtra(BOOKJSONKEY, s);
-
-            username = editText.getText().toString();
-            intent.putExtra(USERNAMEKEY, username);
-            intent.putExtra(FIRSTLOADKEY, true);
-
-            if (running && network) {
-                startActivity(intent);
-            }
-        }
-
-        @Override
-        protected void onProgressUpdate(Integer... percent) {
-            if (running) {
-                super.onProgressUpdate(percent);
-                progressDialog.setMessage(percent[0] * 2 + "%");
-            }
-        }
-
-        @Override
-        protected void onCancelled(String s) {
-            super.onCancelled(s);
-            running = false;
-        }
-
-        private File getTempFile(String name) {
-            File file = null;
-            try {
-                file = new File(getCacheDir(), name);
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-            return file;
-        }
-
-        private void storeFile(String filename, String value) {
-            File tempFile = getTempFile(filename);
-            FileOutputStream outputStream = null;
-            try {
-                outputStream = new FileOutputStream(tempFile);
-                outputStream.write(value.getBytes());
-                Log.d(MAINACTIVITYTAG, "Store cache to disk: " + filename);
-                Log.d(MAINACTIVITYTAG, "Cache File: " + tempFile.getName());
-                outputStream.close();
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-
-        private boolean isTotalNotZero(String json) {
-            if (json != null && !json.isEmpty()) {
-                try {
-                    JSONObject jsonObject = new JSONObject(json);
-                    int total = jsonObject.getInt("total");
-                    if (total > 0) {
-                        return true;
-                    }
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            return false;
-        }
-    }
-}
