+++ demo/RecyclerViewDemo/app/src/main/java/com/example/leopeng/recyclerviewdemo/activity/DetailViewActivity.java
@@ -1,4 +1,4 @@
-package com.example.leopeng.recyclerviewdemo;
+package com.example.leopeng.recyclerviewdemo.activity;
 
 import android.content.Intent;
 import android.graphics.Color;
@@ -13,8 +13,11 @@ import android.widget.ImageView;
 import android.widget.RatingBar;
 import android.widget.TextView;
 
+import com.example.leopeng.recyclerviewdemo.util.ImageDownloaderTask;
+import com.example.leopeng.recyclerviewdemo.R;
+import com.example.leopeng.recyclerviewdemo.model.BookAdapter;
+
 import java.util.List;
-import java.util.Map;
 
 import me.gujun.android.taggroup.TagGroup;
