+++ demo/RecyclerViewDemo/app/src/main/java/com/example/leopeng/recyclerviewdemo/util/Constant.java
@@ -1,0 +1,13 @@
+package com.example.leopeng.recyclerviewdemo.util;
+
+/**
+ * Created by leopeng on 28/02/2017.
+ */
+
+public class Constant {
+    public final static String BOOKJSONKEY = "bookJSON";
+    public final static String USERNAMEKEY = "username";
+    public final static String FIRSTLOADKEY = "firstLoad";
+    public final static String SEARCHKEY = "search";
+    public final static String MAINACTIVITYTAG = "mainActivity";
+}
