+++ demo/RecyclerViewDemo/app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
     android:clickable="true"
-    tools:context="com.example.leopeng.recyclerviewdemo.MainActivity"
+    tools:context="com.example.leopeng.recyclerviewdemo.activity.MainActivity"
     android:id="@+id/backgroundView"
     android:focusableInTouchMode="true">
