+++ demo/RecyclerViewDemo/app/src/main/res/layout/activity_recycler_view.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.leopeng.recyclerviewdemo.MainActivity">
+    tools:context="com.example.leopeng.recyclerviewdemo.activity.MainActivity">
 
     <android.support.v7.widget.Toolbar
         android:layout_width="match_parent"
